Make the majority of the Mapper's "strict" with mapping target attributes
By default, if mapstruct cannot find a target attribute, it emits a
compile time warning (that no one sees) and then continues with not
mapping anything into that attribute.  With this change, (most)
mappers have been configured to make this a compile error.

As a consequence:

 * Some target attributes have been fixed and are now mapped.
 * Most missing target attributes have now been explicitly declared as
   missing (along with a `FIXME` comment).
 * A few mappers are left at the default to close off the scope of this
   change (time constrained).

This will make it much harder for us to overlook mapping of attributes
in the future and will make it less of a "Whack-a-mole" job to
maintain the TO<-> DAO mapping in the BKG/eBL standard.

Signed-off-by: Niels Thykier <nit@globeteam.com>