Double check the drafted transport document
The SI level validator does a decent job at trimming out most of the
issues. However, there are still some cases where the SI level
validator does not catch the problems.

To ensure that we do not generate invalid TDs (as far as we know/can
validate the TO), we now map the draft TD to a TO and run the TO
validation on it. If it finds anything, then we throw `AssertionError`
to stop the draft from happening.  It is not ideal, but it is better
than the reference implementation generating a TD that it would not
accept itself.

Signed-off-by: Niels Thykier <nit@globeteam.com>