DT-480 [4/?]: Make it possible to "lazily" resolve shipments and equipment
This is a stepping stone towards solving DT-480 and later DT-578 by
/enabling/ us to disconnect the POST /shipping-instructions from the
resolution of Shipments and Equipments.  At this time, the resolution
still happens immediately but the database allows for the attributes
be nullable and unresolved without any test breaking.

As a side-effect of this and simpler mapping logic, the
`StuffingService` became 90% obsolete and the remaining 10% was moved
into the `ShippingInstructionService`.

For a few mappers, missing `uses` were added as well to avoid mapping
errors due to changes to (indirect) entities.

Signed-off-by: Niels Thykier <nit@globeteam.com>